Routing code refactored.
